mindbot/backend/requirements.txt
# Web Framework & Server
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0

# Groq, LLM & Environment
groq==1.1.2
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1

# Database & Storage
sqlalchemy==2.0.38
faiss-cpu==1.13.2
numpy==2.4.3

# File Processing
pypdf==5.3.0